+++ Matilda.Test/IntegrationTests/IntegrationTests.cs
@@ -1,71 +1,10 @@
 using Matilda;
 
-namespace MatildaTests;
+namespace MatildaTests.IntegrationTests;
 
 [TestClass]
-public class IntegrationTests
+public class IntegrationTests : IntegrationTestHelper
 {
-    // Helper function to locate the integration test script directory
-    private static readonly string ScriptFolder =
-        Path.GetFullPath(Path.Combine(
-            AppContext.BaseDirectory,
-            "../../../TestMatildaScripts/IntegrationTestsScripts"));
-
-    private Program ParseFile(string fileName)
-    {
-        string path = Path.Combine(ScriptFolder, fileName);
-
-        Assert.IsTrue(File.Exists(path), $"Test script file was not found: {path}");
-
-        Scanner scanner = new Scanner(path);
-        Parser parser = new Parser(scanner);
-
-        parser.Parse();
-
-        Assert.IsFalse(parser.hasErrors(), "Parser had errors.");
-
-        return parser.mainNode;
-    }
-    private void AssertNoTypeErrors(Program program)
-    {
-        EnvVT envVT = new EnvVT();
-        EnvPT envPT = new EnvPT();
-        EnvST envST = new EnvST();
-
-        TypeChecker typeChecker = new TypeChecker(program, envVT, envPT, envST);
-
-        Assert.IsFalse(typeChecker.HasErrors(), $"Type checker has erros: {string.Join("\n", typeChecker.errors)}");
-    }
-
-    private static readonly object ConsoleLock = new object();
-
-    private string RunProgram(Program program)
-    {
-        lock (ConsoleLock)
-        {
-            EnvV envV = new EnvV();
-            EnvP envP = new EnvP();
-            EnvS envS = new EnvS();
-
-            TextWriter originalOutput = Console.Out;
-            StringWriter output = new StringWriter();
-
-            try
-            {
-                Console.SetOut(output);
-
-                Interpreter.EvalTopLevelDeclarations(program.TopLevelDeclarations, envP, envS);
-                Interpreter.EvalStmt(program.Stmt, envV, envP, envS);
-            }
-            finally
-            {
-                Console.SetOut(originalOutput);
-            }
-
-            return output.ToString().Trim();
-        }
-    }
-
     [TestMethod]
     public void FunctionCallProgramBindsExpectedValueInEnvironment()
     {
