+++ Matilda.Test/MatildaCSVFiles/TableDeclarationTest.csv
@@ -1,3 +1,4 @@
-hej,dig
-1,2
-3,4
+Id,name
+1,Alice
+2,Bob
+3,Steve
