--- Matilda.Test/TypeChecker/TestMatildaScriptsTypeChecker/TableDeclaration.csv
@@ -1,4 +1,0 @@
-Id,name
-1,Alice
-2,Bob
-3,Steve
