+++ Matilda.Test/Utils/IntegrationTests.cs
@@ -1,0 +1,13 @@
+using Matilda;
+
+namespace MatildaTests.IntegrationTests;
+
+public class IntegrationTestHelper : TestHelper
+{
+    public override string ScriptFolder
+    {
+        get => Path.GetFullPath(Path.Combine(
+            AppContext.BaseDirectory,
+            "../../../TestMatildaScripts/IntegrationTestsScripts"));
+    }
+}
